auth-service/requirements.txt
fastapi[standard]
uvicorn[standard] 
pydantic 
sqlalchemy
passlib
jinja2
pyjwt